***Func renamed to **Ext
