+++ src/lib/maths.ts
@@ -1,9 +1,10 @@
+export * as MathExt from "./maths/mathExt.js";
 export { default as Matrix } from "./maths/matrix.js";
 export { default as Matrix4 } from "./maths/matrix4.js";
-export * as Matrix4Func from "./maths/matrix4Func.js";
+export * as Matrix4Ext from "./maths/matrix4Ext.js";
 export { default as Rectangle } from "./maths/rectangle.js";
 export { default as Vector2 } from "./maths/vector2.js";
-export * as Vector2Func from "./maths/vector2Func.js";
+export * as Vector2Ext from "./maths/vector2Ext.js";
 export { default as Vector3 } from "./maths/vector3.js";
-export * as Vector3Func from "./maths/vector3Func.js";
+export * as Vector3Ext from "./maths/vector3Ext.js";
 export { default as Vector4 } from "./maths/vector4.js";
