+++ src/lib/maths/mathExt.ts
